Implement quick switching to last tab (closes #14)
Use hold key + space to switch to the last tab navigated to with KeepTabs. Like
alt+tab OS window switching, but with only a history of 1. Doesn't work for
navigation via tab search, and this will take a bit more work to fix, so will
do in a separate future commit.