Track all tab changes for last_tab_id (closes #39)
Use an onActivated listener to update last_tab_id every time the active tab is
changed, not just when it's through KeepTabs. This does break one small
behaivor: before when the user opened tab search with the hotkey, then manually
closed tab search without selecting a tab, the last tab would be saved as the
tab that tab search was launched from. Now, it's saved as the tab search tab
(which has been closed), so trying to navigate to the previous tab doesn't do
anything. But this is a pretty obscure use case (you shouldn't really be
manually cancelling searches often), and it's not even that big a disruption
when it happens, so I'm fine with it.