Implement hold key customization (closes #1)
In the process, I switched over from using e.which and e.keyCode to the
Chrome-specific property e.key, which provides a nice semantic string name for
the key, eliminating the need for coupled pairs of KEY_CODE and SYMBOL (now
replaced with a single KEYVAL) throughout the extension. I also had to move
what was in the keypress handler to the keydown handler in hotkeys.js, since
Control and Alt are treated as modifier keys and no keypress event is triggered
when you press e.g. Ctrl+f.