Use stack for previous tab tracking (closes #44)
While this does reduce and simplify a lot of code, it also gives rise to a
couple of bugs, namely:

1) When the user opens tab search with the hotkey and navigates through it to a
tab in another window, the original window's active tab will then be the
rightmost tab (rather than the tab from which search was launched) after the
search tab closes itself; this also messes up the tab history for navigating to
the previous tab. This is bad, but I'm letting it be for now since we're
planning to implement the over-screen search UI for the v1.0.0 release
milestone anyway and that will eliminate this problem.

2) When the user navigates with tab search via the popup to another tab within
the same window (didn't test cross-window yet), the tab navigated to is not
focused and doesn't receive keystrokes (including hotkeys for KT). I'm not sure
why this happens after this commit and not before, as I can't see any relevant
changes. It's a very small issue so I'm leaving it for now, and we'll file an
issue later if it's still a problem after the new search UI.