Add custom prefix matching for dedup (closes #24)
The hotkey entries in the options now have an additional field, Match Prefix,
that is required only when Deduplicate is checked. Any tab with a URL starting
with the Match Prefix is considered a match to this hotkey for deduplication
purposes, but the Target (which can the same or more specific than Match
Prefix) is the URL that will be opened if no match already exists.

Important note: the Match Prefix has to be usable as a Chrome "URL match
pattern" when just a "*" is added to the end of it, so I added a rudimentary
regex validation on its input element. For some reason (I'm guessing related to
Foundation), the default popups aren't appearing when the user tries to save
with an invalid Match Prefix, but the page simply focuses the first box with an
invalid input, which is very unclear. Much clearer notification of what's
invalid and how to fix it should be a part of #20.

Also, this commit removes the legacy code for loading and converting the old
format of options, as planned. Options will not load correctly for any user who
has not first saved their options with the previous commit's code. (But they
can just check out the previous commit, open and save their options, then
checkout this commit again.)