Add unsaved options warning on close (part of #26)
Unfortunately, there's no way to give the warning when the user closes the
window in general, but we can at least do it when they use our "Close" button.
The button also displays when there are unsaved changes with different text. I
want to get the save and close buttons to float and always stay on screen, but
I can't get position: fixed to work within the extension options page. Since
we're going to be moving the options to the popup anyway (#19), I'll try it
then.