Mark unsaved when hold key choice is changed
This was working a few commits back, but I deleted the line that added the
listener to all INPUTTABLE elements, thinking it was unnecessary because such
listeners are already added to each row as they're added. I forgot we still
need to add the listener to the hold key choice <select> element.