Make CSS KT-specific to not collide with websites
There are at least a few websites I noticed problems on, one of which was
articles (not the main page) on androidauthority.com. This is a quick fix, a
bit ugly to add #keeptabs_overlay before everything. We should move to Sass or
Less at some point anyway, and then our CSS will be cleaner.