Remove unnecessary chrome.tabs.query calls
The one in createNewTab is definitely doing nothing; the tabs parameter of the
callback isn't used. I'm 95% sure changing from the chrome.tabs.query to using
the current tab from tab history will not change the behavior in
closeCurrentTab, but we'll find out soon enough if there are edge cases where
it does.